Fix output paths for `setup-fhenix`
The output paths were mistakenly using the old output paths from the
`install` step instead of the `output` step -- which ultimately means
the paths were not normalized for the correct operating system (e.g.
on windows, the path would appear as `/d/...` instead of `D:\...`).

This updates the output paths to point to the correct ones.